Update from upstream repo orientechnologies/orientdb@master

+++ lucene/src/test/java/com/orientechnologies/lucene/test/LuceneBackupRestoreTest.java
@@ -26,18 +26,22 @@ import com.orientechnologies.orient.core.sql.OCommandSQL;
 import com.orientechnologies.orient.core.sql.query.OSQLSynchQuery;
 import org.junit.*;
 import org.junit.rules.TemporaryFolder;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
+import java.util.Locale;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Created by Enrico Risa on 07/07/15.
  */
+@RunWith(JUnit4.class)
 public class LuceneBackupRestoreTest {
 
   @Rule
@@ -47,6 +51,8 @@ public class LuceneBackupRestoreTest {
 
   @Before
   public void setUp() throws Exception {
+    final String os = System.getProperty("os.name").toLowerCase(Locale.ENGLISH);
+    Assume.assumeFalse(os.contains("win"));
 
     String url = "plocal:./target/" + getClass().getName();
 
@@ -75,7 +81,9 @@ public class LuceneBackupRestoreTest {
 
   @After
   public void tearDown() throws Exception {
-    dropIfExists();
+    final String os = System.getProperty("os.name").toLowerCase(Locale.ENGLISH);
+    if (!os.contains("win"))
+      dropIfExists();
 
   }
